fix: ex01

+++ ex01/include/Fixed.hpp
@@ -6,7 +6,7 @@
 /*   By: takawauc <takawauc@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/15 12:40:30 by takawauc          #+#    #+#             */
-/*   Updated: 2025/12/15 14:18:19 by takawauc         ###   ########.fr       */
+/*   Updated: 2026/02/03 21:32:36 by takawauc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,13 +15,14 @@
 
 #include <iostream>
 
-class Fixed {
+class Fixed
+{
 public:
   Fixed(void);
   Fixed(int integer);
-  Fixed(double float_num);
-  Fixed(const Fixed &other);
-  Fixed &operator=(const Fixed &other);
+  Fixed(float float_num);
+  Fixed(const Fixed& other);
+  Fixed& operator=(const Fixed& other);
   ~Fixed(void);
   int getRawBits(void) const;
   void setRawBits(int const raw);
@@ -34,6 +35,6 @@ private:
   static const int _frac_bits = 8;
 };
 
-std::ostream &operator<<(std::ostream &os, const Fixed &c);
+std::ostream& operator<<(std::ostream& os, const Fixed& c);
 
 #endif
